% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeated_irregular_poly.R
\name{ripoly}
\alias{ripoly}
\title{Handles polynomials for repeated irregular measures.
Used in longitudinal data modeling ex: with lmer, lcmm, predict effects packages instead of poly function}
\usage{
ripoly(x, ..., degree = 1, coefs = NULL, raw = FALSE, simple = FALSE)
}
\arguments{
\item{x}{Numeric vector or matrix with a single column}

\item{degree}{Polynomial order}

\item{coefs}{Coeffecients of previous fit}

\item{raw}{T for raw data, F for orthogonal polynomials}

\item{simple}{T for matrix result, F for poly object}
}
\description{
Handles polynomials for repeated irregular measures.
Used in longitudinal data modeling ex: with lmer, lcmm, predict effects packages instead of poly function
}
\examples{
# General example
ripoly(c(1:5), 2)

# Linear mixed model example
model1 <- lmer(Reaction ~ ripoly(Days, 2) + (ripoly(Days, 2) | Subject), sleepstudy))

# Predictor effects package
ef <- predictorEffects(model1, Days)
}
